Move printOutput helper
