+++ src/lib/helpers/print-output.js
